Rename package and methods to proper names
