--- app/src/main/java/com/example/appName/data/repository/dummyuser/ExampleUserModule.kt
@@ -1,13 +1,0 @@
-package com.example.appName.data.repository.dummyuser
-
-import dagger.Module
-import dagger.Provides
-import javax.inject.Singleton
-
-@Module
-class ExampleUserModule {
-
-    @Provides
-    @Singleton
-    fun provideDummyUserRepository(): ExampleUserRepository = LocalExampleUserRepository()
-}
