+++ app/src/main/java/com/example/appName/data/repository/exampleuser/ExampleUserModule.kt
@@ -1,0 +1,13 @@
+package com.example.appName.data.repository.exampleuser
+
+import dagger.Module
+import dagger.Provides
+import javax.inject.Singleton
+
+@Module
+class ExampleUserModule {
+
+    @Provides
+    @Singleton
+    fun provideExampleUserRepository(): ExampleUserRepository = LocalExampleUserRepository()
+}
