+++ app/src/main/java/com/example/appName/data/repository/exampleuser/LocalExampleUserRepository.kt
@@ -1,4 +1,4 @@
-package com.example.appName.data.repository.dummyuser
+package com.example.appName.data.repository.exampleuser
 
 import com.example.appName.data.model.ExampleUser
 import io.reactivex.Single
